chore: move fading logic into separate hook
